Move ts-eslint into the utils package
