+++ packages/eslint-plugin/src/rules/ban-types.ts
@@ -1,5 +1,4 @@
-import { TSESTree, AST_NODE_TYPES } from '@typescript-eslint/util';
-import { ReportFixFunction } from 'ts-eslint';
+import { TSESLint, TSESTree, AST_NODE_TYPES } from '@typescript-eslint/util';
 import * as util from '../util';
 
 type Options = [
@@ -94,7 +93,7 @@ export default util.createRule<Options, MessageIds>({
             let customMessage = '';
             const bannedCfgValue = bannedTypes[node.name];
 
-            let fix: ReportFixFunction | null = null;
+            let fix: TSESLint.ReportFixFunction | null = null;
 
             if (typeof bannedCfgValue === 'string') {
               customMessage += ` ${bannedCfgValue}`;
