+++ packages/eslint-plugin/src/rules/prefer-interface.ts
@@ -1,5 +1,4 @@
-import { TSESTree } from '@typescript-eslint/util';
-import { RuleFix } from 'ts-eslint';
+import { TSESLint, TSESTree } from '@typescript-eslint/util';
 import * as util from '../util';
 
 export default util.createRule({
@@ -33,7 +32,7 @@ export default util.createRule({
           messageId: 'interfaceOverType',
           fix(fixer) {
             const typeNode = node.typeParameters || node.id;
-            const fixes: RuleFix[] = [];
+            const fixes: TSESLint.RuleFix[] = [];
 
             const firstToken = sourceCode.getFirstToken(node);
             if (firstToken) {
