+++ packages/eslint-plugin/src/rules/require-array-sort-compare.ts
@@ -1,5 +1,5 @@
-import * as ts from 'typescript';
 import { TSESTree } from '@typescript-eslint/util';
+import ts from 'typescript';
 import * as util from '../util';
 
 export default util.createRule({
