+++ packages/eslint-plugin/src/rules/type-annotation-spacing.ts
@@ -1,5 +1,5 @@
-import * as util from '../util';
 import { TSESTree } from '@typescript-eslint/util';
+import * as util from '../util';
 
 type Options = [
   {
