+++ packages/eslint-plugin/src/util/createRule.ts
@@ -1,9 +1,4 @@
-import RuleModule, {
-  RuleListener,
-  RuleMetaData,
-  RuleMetaDataDocs,
-  RuleContext,
-} from 'ts-eslint';
+import { TSESLint } from '@typescript-eslint/util';
 import { applyDefault } from './applyDefault';
 
 // note - cannot migrate this to an import statement because it will make TSC copy the package.json to the dist folder
@@ -16,19 +11,19 @@ type RemoveProps<
 > = Pick<TObj, Exclude<keyof TObj, TKeys>>;
 
 // we'll automatically add the url + tslint description for people.
-type CreateRuleMetaDocs = RemoveProps<RuleMetaDataDocs, 'url'> & {
+type CreateRuleMetaDocs = RemoveProps<TSESLint.RuleMetaDataDocs, 'url'> & {
   tslintName?: string;
 };
 type CreateRuleMeta<TMessageIds extends string> = {
   docs: CreateRuleMetaDocs;
-} & RemoveProps<RuleMetaData<TMessageIds>, 'docs'>;
+} & RemoveProps<TSESLint.RuleMetaData<TMessageIds>, 'docs'>;
 
 // This function will get much easier to call when this is merged https://github.com/Microsoft/TypeScript/pull/26349
 // TODO - when the above rule lands; add type checking for the context.report `data` property
 export function createRule<
   TOptions extends Readonly<any[]>,
   TMessageIds extends string,
-  TRuleListener extends RuleListener = RuleListener
+  TRuleListener extends TSESLint.RuleListener = TSESLint.RuleListener
 >({
   name,
   meta,
@@ -39,10 +34,10 @@ export function createRule<
   meta: CreateRuleMeta<TMessageIds>;
   defaultOptions: TOptions;
   create: (
-    context: RuleContext<TMessageIds, TOptions>,
+    context: TSESLint.RuleContext<TMessageIds, TOptions>,
     optionsWithDefault: TOptions,
   ) => TRuleListener;
-}): RuleModule<TMessageIds, TOptions, TRuleListener> {
+}): TSESLint.RuleModule<TMessageIds, TOptions, TRuleListener> {
   return {
     meta: {
       ...meta,
