+++ packages/eslint-plugin/src/util/misc.ts
@@ -2,9 +2,7 @@
  * @fileoverview Really small utility functions that didn't deserve their own files
  */
 
-import { TSESTree, AST_NODE_TYPES } from '@typescript-eslint/util';
-import RuleModule from 'ts-eslint';
-import { SourceCode } from 'ts-eslint';
+import { AST_NODE_TYPES, TSESLint, TSESTree } from '@typescript-eslint/util';
 
 /**
  * Check if the context file name is *.ts or *.tsx
@@ -27,7 +25,7 @@ export function upperCaseFirst(str: string) {
   return str[0].toUpperCase() + str.slice(1);
 }
 
-type InferOptionsTypeFromRuleNever<T> = T extends RuleModule<
+type InferOptionsTypeFromRuleNever<T> = T extends TSESLint.RuleModule<
   never,
   infer TOptions
 >
@@ -36,7 +34,7 @@ type InferOptionsTypeFromRuleNever<T> = T extends RuleModule<
 /**
  * Uses type inference to fetch the TOptions type from the given RuleModule
  */
-export type InferOptionsTypeFromRule<T> = T extends RuleModule<
+export type InferOptionsTypeFromRule<T> = T extends TSESLint.RuleModule<
   any,
   infer TOptions
 >
@@ -46,7 +44,7 @@ export type InferOptionsTypeFromRule<T> = T extends RuleModule<
 /**
  * Uses type inference to fetch the TMessageIds type from the given RuleModule
  */
-export type InferMessageIdsTypeFromRule<T> = T extends RuleModule<
+export type InferMessageIdsTypeFromRule<T> = T extends TSESLint.RuleModule<
   infer TMessageIds,
   any
 >
@@ -71,7 +69,7 @@ export function getNameFromPropertyName(
  */
 export function getNameFromClassMember(
   methodDefinition: TSESTree.MethodDefinition | TSESTree.ClassProperty,
-  sourceCode: SourceCode,
+  sourceCode: TSESLint.SourceCode,
 ): string {
   if (keyCanBeReadAsPropertyName(methodDefinition.key)) {
     return getNameFromPropertyName(methodDefinition.key);
