+++ packages/parser/src/scope/scope-manager.ts
@@ -1,4 +1,4 @@
-import { TSESTree } from '@typescript-eslint/util';
+import { TSESTree } from '@typescript-eslint/typescript-estree';
 import EslintScopeManager, {
   ScopeManagerOptions,
 } from 'eslint-scope/lib/scope-manager';
