+++ packages/parser/tests/lib/visitor-keys.ts
@@ -1,4 +1,4 @@
-import { AST_NODE_TYPES } from '@typescript-eslint/util';
+import { AST_NODE_TYPES } from '@typescript-eslint/typescript-estree';
 import { visitorKeys } from '../../src/visitor-keys';
 
 //------------------------------------------------------------------------------
