+++ packages/typescript-estree/package.json
@@ -37,8 +37,7 @@
   },
   "dependencies": {
     "lodash.unescape": "4.0.1",
-    "semver": "5.5.0",
-    "@typescript-eslint/util": "1.4.2"
+    "semver": "5.5.0"
   },
   "peerDependencies": {
     "typescript": "*"
