+++ packages/typescript-estree/src/ts-estree/index.ts
@@ -1,0 +1,5 @@
+import * as TSESTree from './ts-estree';
+
+export { TSESTree };
+export * from './ast-node-types';
+export * from './ts-nodes';
