+++ packages/typescript-estree/src/ts-estree/ts-nodes.ts
