+++ packages/util/src/index.ts
@@ -1,5 +1,14 @@
-import * as TSESTree from './ts-estree';
+import * as TSESLint from './ts-eslint';
 
-export * from './ast-node-types';
-export * from './ts-nodes';
-export { TSESTree };
+export { TSESLint };
+
+// for convenience's sake - export the types directly from here so consumers
+// don't need to reference/install both packages in their code
+export {
+  AST_NODE_TYPES,
+  AST_TOKEN_TYPES,
+  ParserOptions,
+  ParserServices,
+  TSESTree,
+  TSNode,
+} from '@typescript-eslint/typescript-estree';
