+++ packages/util/src/ts-eslint/ParserOptions.ts
@@ -1,0 +1,21 @@
+export interface ParserOptions {
+  loc?: boolean;
+  comment?: boolean;
+  range?: boolean;
+  tokens?: boolean;
+  sourceType?: 'script' | 'module';
+  ecmaVersion?: number;
+  ecmaFeatures?: {
+    globalReturn?: boolean;
+    jsx?: boolean;
+  };
+  // ts-estree specific
+  filePath?: string;
+  project?: string | string[];
+  useJSXTextNode?: boolean;
+  errorOnUnknownASTType?: boolean;
+  errorOnTypeScriptSyntacticAndSemanticIssues?: boolean;
+  tsconfigRootDir?: string;
+  extraFileExtensions?: string[];
+  warnOnUnsupportedTypeScriptVersion?: boolean;
+}
