+++ packages/util/tsconfig.build.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "./dist",
+    "rootDir": "./src",
     "resolveJsonModule": true
   },
   "include": [
