+++ packages/util/tsconfig.json
@@ -3,8 +3,6 @@
   "include": [
     "src",
     "typings",
-    // include the parser's ambient typings because the parser exports them in its type defs
-    "../parser/typings",
     "tests",
     "tools"
   ]
